long-edge/backend/requirements.txt
# Momentum Hunter - Python Dependencies

# Trading & Market Data
alpaca-py>=0.23.0
alpaca-trade-api>=3.2.0

# AI & ML
anthropic>=0.18.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Caching
redis>=5.0.1

# Web Framework (for API/Dashboard backend)
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.2

# Logging & Monitoring
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
